Introduce Labrador Coding Standard
- Implements a customized subset of the PSR2 PHP_CodeSniffer ruleset to
provide a consistent format for which all Labrador code should adhere
to. Primarily the differences involve the use of Allman style bracing in
favor of the OTBS.

- Provides labrador-cs and labrador-cbf binaries to ensure that your
phpcs and phpcbf commands operate with the appropriate standard. When
working with Labrador packages it is HIGHLY recommended that you use
these commands instead of phpcs directly as the default PSR-2 style is
NOT what we code with.